fix fieldSelector cannot work
